+++ test/clj/metosin/dates/generators_test.clj
@@ -1,3 +1,1 @@
-(ns metosin.dates.generators-test
-  (:require [metosin.dates.generators :refer :all]
-            [clojure.test :refer :all]))
+(ns metosin.dates.generators-test)