+++ test/clj/metosin/email_test.clj
@@ -1,3 +1,1 @@
-(ns metosin.email-test
-  (:require  [metosin.email :refer :all]
-             [clojure.test :refer :all]))
+(ns metosin.email-test)