+++ test/clj/metosin/jdbc/joda/time_test.clj
@@ -1,6 +1,6 @@
 (ns metosin.jdbc.joda.time-test
-  (:require [metosin.jdbc :refer :all]
-            [clojure.test :refer :all]
+  (:require [clojure.java.jdbc :as jdbc]
+            [clojure.test :refer [deftest testing is]]
             [metosin.dates :as dates]
             metosin.jdbc.joda.time)
   (:import [org.joda.time DateTime LocalDate]))
@@ -10,23 +10,22 @@
               :subname "db/metosin-common-test"})
 
 (deftest metosin-jdbc-test
-  (with-db-connection [db h2-spec]
-    (db-do-commands db
-      (create-table-ddl :test_table
-                        [[:id "bigint primary key auto_increment"]
-                         [:datetime "timestamp"]
-                         [:date "date"]]))
+  (jdbc/with-db-connection [db h2-spec]
+    (jdbc/db-do-commands db
+                         (jdbc/create-table-ddl :test_table
+                                                [[:id "bigint primary key auto_increment"]
+                                                 [:datetime "timestamp"]
+                                                 [:date "date"]]))
 
-    (insert! db :test_table {:datetime (dates/date-time 2015 2 9 13 1)
-                             :date (dates/date 2015 2 9)})
+    (jdbc/insert! db :test_table {:datetime (dates/date-time 2015 2 9 13 1)
+                                  :date (dates/date 2015 2 9)})
 
     (testing "read inserted values"
-      (let [{:keys [datetime datetime-default date date-default]}
-            (first (query db ["SELECT * from test_table"]))]
+      (let [{:keys [datetime date]}
+            (first (jdbc/query db ["SELECT * from test_table"]))]
 
         (is (= (dates/date-time 2015 2 9 13 1)
                datetime))
 
         (is (= (dates/date 2015 2 9)
-               date))
-        ))))
+               date))))))