+++ test/clj/metosin/postgres/types_test.clj
@@ -1,13 +1,10 @@
 (ns metosin.postgres.types-test
-  (:require [metosin.postgres.types :refer :all]
-            [clojure.java.jdbc :as jdbc]
-            [clojure.test :refer :all]))
-
+  (:require [metosin.postgres.types :as t]
+            [clojure.test :refer [deftest testing is]]))
 
 (deftest pgobject->clj-test
   (testing "json"
-    (is (= {:foo "a"} (pgobject->clj (->PGobject "json" "{\"foo\":\"a\"}"))))
-    (is (= {:foo "a"} (pgobject->clj (write-json {:foo "a"}))))
-    (is (= ["a" "b"] (pgobject->clj (->PGobject "json" "[\"a\",\"b\"]"))))
-    (is (= ["a" "b"] (pgobject->clj (write-json ["a" "b"]))))
-    ))
+    (is (= {:foo "a"} (t/pgobject->clj (t/->PGobject "json" "{\"foo\":\"a\"}"))))
+    (is (= {:foo "a"} (t/pgobject->clj (t/write-json {:foo "a"}))))
+    (is (= ["a" "b"] (t/pgobject->clj (t/->PGobject "json" "[\"a\",\"b\"]"))))
+    (is (= ["a" "b"] (t/pgobject->clj (t/write-json ["a" "b"]))))))