+++ test/clj/metosin/ring/util/etag_test.clj
@@ -1,14 +1,12 @@
 (ns metosin.ring.util.etag-test
-  (:require [metosin.ring.util.etag :refer :all]
-            [clojure.java.io :as io]
-            [clojure.test :refer :all]
-            [metosin.dates :as d]))
+  (:require [metosin.ring.util.etag :as etag]
+            [clojure.test :refer [deftest is testing]]))
 
 (def req {:status 200, :body nil, :headers {"if-none-match" "a"}})
 
 (deftest etag-test
   (testing "same etag"
-    (is (true? (none-match? req "a"))))
+    (is (true? (etag/none-match? req "a"))))
 
   (testing "different etag"
-    (is (not (none-match? req "b")))))
+    (is (not (etag/none-match? req "b")))))