+++ test/clj/metosin/sql_test.clj
@@ -1,3 +1,1 @@
-(ns metosin.sql-test
-  (:require [metosin.sql :refer :all]
-            [clojure.test :refer :all]))
+(ns metosin.sql-test)